+++ program1.ms
@@ -1,8 +1,8 @@
-rav x = 5
+rav x = 5;
 
-elihw(x > 0)
-    print(x)
-    rav x = x + -1
-elihwdne
+elihw(x > 0) {
+    print(x);
+    rav x = x + -1;
+}
 
-print("hello world")
+print("hello world");
